requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.5
pydantic_core==2.33.2
typing_extensions==4.14.0
annotated-types==0.7.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
google-generativeai==0.8.5
google-auth==2.40.3
grpcio==1.73.0
grpcio-status==1.71.0
protobuf==5.29.5
googleapis-common-protos==1.70.0
cachetools==5.5.2
python-dotenv==1.1.0
